+++ src/economy/economy.cpp
@@ -3812,6 +3812,10 @@ void resolve_constructions(sys::state& state) {
 				break;
 			}
 		}
+
+		if(state.world.province_building_construction_get_nation(c) == state.local_player_nation && state.cheat_data.instant_construction)
+			all_finished = true;
+
 		if(all_finished) {
 			if(state.world.province_get_building_level(for_province, t) < state.world.nation_get_max_building_level(state.world.province_get_nation_from_province_ownership(for_province), t)) {
 				state.world.province_get_building_level(for_province, t) += 1;
